XIVY-16170 Set write permissions on the root folder `/ivy` for the group
Seems like `COPY` does also copy the permissions of files, as
explained in the Dockerfile specs. But the copied root folder
won't be copied, it will be created by COPY itself.

No longer download in `/tmp` directory in staging layer, to
make the `Dockerfile` more readable.

Remove the `ARG=IVY_HOME`, not crucial in this `Dockerfile`.